+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## v4.2.1 (unreleased)
 
 * Change the `volumes` subcommand to handle sizes in GB.
+* Add `--checks-grace-period` to set the grace period for health checks, for example with `koyeb service update app/service --checks 8000:http:/healtcheck --checks-grace-period 8000=10`.
 
 ## v4.2.0